cast numbe of periods to int in model.py

+++ site_forecast_app/models/pvnet/model.py
@@ -281,7 +281,7 @@ class PVNetModel:
 
         forecast_timesteps = pd.date_range(
             start=self.t0 - pd.Timedelta("52h"),
-            periods=4 * 24 * 4.5,
+            periods=int(4 * 24 * 4.5),
             freq="15min",
         )
